Fix retention.ms long value

+++ src/LEGO.AsyncAPI.Bindings/Kafka/KafkaChannelBinding.cs
@@ -47,7 +47,7 @@ namespace LEGO.AsyncAPI.Bindings.Kafka
         private static FixedFieldMap<TopicConfigurationObject> kafkaChannelTopicConfigurationObjectFixedFields = new ()
         {
             { "cleanup.policy", (a, n) => { a.CleanupPolicy = n.CreateSimpleList(s => s.GetScalarValue()); } },
-            { "retention.ms", (a, n) => { a.RetentionMilliseconds = n.GetIntegerValue(); } },
+            { "retention.ms", (a, n) => { a.RetentionMilliseconds = n.GetLongValue(); } },
             { "retention.bytes", (a, n) => { a.RetentionBytes = n.GetIntegerValue(); } },
             { "delete.retention.ms", (a, n) => { a.DeleteRetentionMilliseconds = n.GetIntegerValue(); } },
             { "max.message.bytes", (a, n) => { a.MaxMessageBytes = n.GetIntegerValue(); } },
